if we flip the axis limits, it gives the unflipped data and axis increasing from right to left. looks good

+++ python/src/scipp/plot/tools.py
@@ -103,7 +103,7 @@ def parse_params(params=None,
 
 
 def make_fake_coord(dim, size, unit=None):
-    args = {"values": np.arange(size)}
+    args = {"values": np.arange(size, dtype=np.float64)}
     if unit is not None:
         args["unit"] = unit
     return sc.Variable(dims=[dim], **args)
